refactor: Move code in the server::server module to the root
This gets rid of the redundant server::server imports.
